Added pytext and utility packages
pytext will be used for processing input from crawler. It will use text proccessing
frameworks example Spacy which is also powerful. utility will be for modules that
that are used by the other packages. section.py and metadata.py in crawler can both be part
of utility package as they are shared by all packages.