Updated crawler text classes to implement utility classes
Section in crawler.text was renamed to Text_Section and it implements Section
of utility package. Text class implement Container class of utility package.
metadata.py of crawler.text was removed in favour of the one in utility
package.

Some improvements were made to the classes and tests updated to
reflect the new changes. metadata.py and test_metadata.py lost their
version history, they were supposed to be moved to utility
instead of copied. But it does not matter as they do not have a long a
long history of changes.