Renamed fetch.py to master_fetch.py
Both the module and the class was renamed with class renamed to Master_Fetch.
Master_Fetch is suitable as it shows that the class builds on top of the other
fetch classes.

register_fetch_class() method was also added for registering fetch classes
dynamically instead of physically writting to the class. This would make
it easier to register fetch classes.