Added tests for Master_Fetch class
The tests are for Master_Fetch class of crawler.fetch package in
master_fetch.py module. There are issues with unclosed files when
running the tests. This happens when method of Master_Fetch class
creates instance of fetch class.

When creating instance of fetch class, file get opened even if source
may not refer to file. Destructor for fetch classes for close the
files when object is destroyed through __del__() method. This
issues had to do with fetch classes as they do not close the
files they open when they get destroyed.

Fetch_Base would need to call .close() when file is destroyed.