Removed versions names and added comments(requirements.txt)
The packages in the requirements.txt are from datanavy search project. They
are not guaranteed to be used in the this project. Version name for
packages were removed as they were specific for requirments of
datanavy search project.

Comments and spacing were added to make it easy to read the file. Its easy
to understand what each package will be used for instead of being
packed together without new lines.