Added html_parse.py and pdf_parse.py files
The two files contain HTML_Parse and PDF_Parse classes which parses
HTML and PDF files. HTML_Parse uses BeautifulSoup and PDF_Parse
uses pdfminer. Both of them are capable of extracting text and
HTML which is supported by the framesworks used.

Title of PDF cannot be found in PDF due to nature of PDF unlike HTML
which has title tags. Both BeautifulSoup and PDF_Parse seems not to
support streamed proccessing. They only build whole document and would
throw error if portion of it is given at a time.

They do accept file like objects but they will create full document
object from it which use more RAM if output is so large. It will be  waste
of memory if BeautifulSoup or pdfminer reads all contents of passed file
fully. That would be a bad practice as pdf files can very large which
will use most of computer RAM.

I dont know if PDFParser of pdfminer really fetches all pdf objects
into memory at once or they do it in more streamed way.

From PDFDocument doctring '''Since a PDF file can be very big, normally it is not loaded at
once. So PDF document has to cooperate with a PDF parser in order to
dynamically import the data as processing goes.'''. Its clear that PDFParser does not load
everythong at once but do it on request when  PDFDocument object requests certain object of
pdf file.

I cant determine it for BeautifulSoup but Python HTML parser included with has
seek() method which can parse html by section(streamed). BeautifulSoup does
use it to create nodes tree.

it seems to parse by potions as it has lines _feed() with statrement 'self.builder.feed(self.markup)'.
Not just that, 'self.builder' has methods like .reset() which are similar to the one
in Python built in parser. _feed() is also called multiple times which shows that html is parsed
as portions. theres is also comment in ._feed() method "# Close out any unfinished strings and close all the open tags.".
Unfished strings and unclosed tags mean tags in which their closing tags have not been supplied.

Remember that <HTML> element will be closed at the end and its possible that the end tag not be provided. Beutifulsoup
parses HTML as line by line or by tag as it has interface similar to that of Python parser which
calls method when it encounter start_tag, data, end_tag, comment which also Beutifulsoup does.

That was long, low level and hard to think of. I wonder how people are able to write such a long code
in their libraries.