Added priority class variable to parse classes
priority variable helps in deciding which parse class to use
if source is valid for multiple parse classes. Its rare for
it to happen but its possible. HTML files are valid for both
Text_Parse and HTML_Parse. Both .html and .txt are text files
that can be are valid for Text_Parse which is not strict.

html files are expected to be parsed with HTML_Parse class not
Text_Parse. priority variable helps determine which parse class
has more priority. The lower prority class the more priority the class
has. priority value for Text_Parse was increased so that HTML_Parse
can be used instead of Text_Parse.

Other text files including .txt, .py, .csv will be parsed with
Text_Parse which just reads the data as it is. If theres a parser
for CSV files, it can be registered and it will be selected in favour
instead of Text_Parse just like html files.