Added functions for adding project source/ to system path
Its not worth it to import 'src' or 'source' as packages to access
'pynavy' package with project modules. project source directory is
added to path when initialising 'tests' package.

I dont know if it will work, but it should work. When running tests with
unittest, the __init__.py of tests will be executed which will add the tests.
Tried to locate unittest discover to pynavy package but it throws errors. Both
tests and their source needs to be accessible to top-level diretcory for tests
to run.

This forces, tests to import folders like 'src' and 'source' as packages as 'pynavy'
cannot be accessed. Solution is to ensure that package is on path so that it
could be accessed by import statements.

Not sure if it will work, it should.