Replaced occurrances of 'navil.crawler' with 'navil'
Commit is part of making crawler part of pynavy independent with
its own repository. 'navil.crawler' is result of find and replace
of 'pynavy.crawler'. navil/ folder is now in projetc source directory
which is directly accessible to test modules.

All tests passed which means the repo is now working.