Fixed few bugs and issues
Changes in this commit are results of tests. The changes were
driven by tests which will be commited in the next commit.

One fix was to supply binary file instead of text in html_to_file()
of PDF_Parse. pdfminer seems to be accepting only binary for
output of html. Memory binary instance was created which passed
as output for html. The binary file contents then gets written copied
to html_file.

That adds extra layer of writting and reading to file. But issue is
that pdf files may contain images which may need to be displayed in
the file. Without binary file, all the images will be lost and replaced
by some odd characters.

I think its better to make html_file binary as html can contain images
which are common in html files. Thats why pdfminer raises errorrs
when passing text file for html output.