Fixed issue with failing test
One test was failing when running all tests in project but when running
the moduel or test class the was running well. It had to do with HTML_Parse
being deregistered. It may have been cleared by other tests espcially
tests that test registering of parse classes.

Unittest maybe running tests conccurrently. But the issue was resolved by registering
the tests on setUp() method which would register the tests everytime test method is tested.