Fixed Fetch class, updated tests
Fetch class was not reacting to source_locates_data argument. Supporting
the argumnet meant having to implement methods of Fetch_Base using Master_Parse class.
Tests did not identify the issue bacause the tests did not test the argumnet.

source_locates_data argument has complicated the project. But that would
make it easy for users when using the high level api such as extract_text().
There seems to be other issues with fetch subpackage and its classes.

Tests for fetch classes needs to be rewritten like that of parse tests.
This issue was identified while creating examples, but that has complicated
Fetch class.

Master_Fetch and Fetch class needs to be reworked.