Updated examples(fetching data)
Added example of using 'source_locates_data' argument.
Seeking of fetch object file before reading was removed
as fetch_obj.read() already seeks to begining of file before reading.

It took a lot of effords to end up making it possible. Lot of
fixes to Fetch class had to be done to end up able complete
this example. The changes are small but it meant having to go
through lot of file and updating tests.