Fixed .is_local_file(source) issues
The regex pattern was not working for windows paths. It was
based on linux paths. It was updated by using os module
and then comparing the contenets based on what makes valid
file path.

The best way to check if file path is valid is creating a file
or directory and then hadle any exceptions and then delete the file
if it was created. is_local_file() just tries to guess if source
may be a file.

I can tell you that there may be lot of issues in the function that
even the tests dont even touch.