Added examples for subclassing Fetch_Base
Creating sub classes of Fetch_Base class is an advantage as
it allows fetch data from anywhere. Data could be fetched from
file system, web, emails or anywhere.

It takes implemeneting minimum of 2 methods to end up creating subclass
of Fetch_Base. Fetch class could also be subclassed but its not worth
it as it depends on Master_Fetch. Unless one does not intend to use
that subclass with Master_Fetch or registering it.

The example classes I think are enough to get someone started.