Updated README.md
Details about the project was added including even examples. Its not
complete but altleat its something start with. More details will be added
if neccessary but at moment I have nowthing to add.

Realise that 'naval' is installed with with 'navaly' when using pip. This is
because theres is already a project called 'naval' in pypi. The project
may need a rename to avoild different name used in imports and when installing.

Maybe renaming 'naval' to 'navaly' is a better choice.