Fixed errors on Windows(Closes #1)
Closes #1

Tests were failing and errors raised when running the project on windows
but on Ubuntu it was running well. Most of issues has to do with implementation
of temporary files. Some issues were also solved and tests for fetch classes
were improved. The total number of tests has increased including even the
time it takes to run them.

This commit is made up of about 24 file changes. Most of this changes are on
parse and fetch subpackages and most of them were all related to file objects.
It was the file objects that caused issues on Windows in the first place thats
why most changes are bsed on them.

The changes are so many but thankfully high level functions like extract_text()
were not influenced. Its not guaranteed if they are not influeced but it seems
so as tests are not failing and examples are running successfully. Some of
the chages may introduse issues and some of them only passed tests by just
a luck.

There may be a lot of issues hidding that even tests never even touched. I just
wonder if those issues will ever show up in the high level functions. There
are some issues that were just left like use of __del__() for clearing resources
which should be replaced by context managers like 'with' statement.

Goal at moment is to see the project working for both windows and linux, also
be available on pypi. Other issues will be resolved later after the project has
been uploaded to pypy.

One problem is that the changes made in this commit were tested on Windows. Its
guaranteed if things will work well for Linux just as I was working on Linux
just to find out there are lot of issues on Windows.

I hope it works on Ubuntu and if doesnt atleat be just minor issues.